backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
